pyproject.toml
[tool.poetry]
authors = ["Shrikanth Upadhayaya <shrik450@gmail.com>"]
description = "AI Query Optimizer"
name = "aqo"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
litellm = "^0.13.2"
mysql-connector-python = "^8.2.0"
psycopg2 = "^2.9.9"
python = "^3.11"
tabulate = "^0.9.0"

[tool.poetry.scripts]
aqo = "aqo.cli:main"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
